requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pywinrm==0.4.3
cryptography==42.0.8
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.3
